Optimize inventory filtering
Previously, when search terms changed, we added a .hidden class to all items, and then selected items that match the search terms using multiple fuzzy attribute selectors and removed the hidden class on them and triggered scroll event on all of them to start lazy loading.

This was very inefficient so we changed it to locating the items manually and only changing the classList where necessary.