Switch to a mediacache instead of maintaining two different fetchers + paper_items.json
We already have mediacache.fcp.cz so no point in duplicating it.

Ideally, we would always fetch data from media server over HTTPs (even in composed)
but to optimize it a bit, we break the encapsulation using `MEDIA_SERVER_LOCAL_DIRECTORY`.

For convenience, we provide fcp’s implementation of mediacache.