Use canvas elements for rendering the clothing layers
We already used them for determining what to undress so we can just use them for view as well.

Primary motivation is making the penguin resizable – previosly, scaling the image would create a mismatch between the TransparencyChecker and the Image coordinates.

As a side-effect, this also results in a much faster switching since we no longer have to drop and re-create `<img>` elements every time an item slot changes.